// matterport/src/models/actionTypes.ts
// export const RECEIVE_FLOOR = "RECEIVE_FLOOR";
export enum MODEL_ACTION {
  RECEIVE_MODELS = "RECEIVE_MODELS",
}

export enum FLOOR_ACTION {
  RECEIVE_FLOOR = "RECEIVE_FLOOR",
}

export const RECEIVE_PLANS = "RECEIVE_PLANS";

// Toast
export enum TOAST_ACTIONS {
  TOAST_ADD = "TOAST_ADD",
  TOAST_REMOVE = "TOAST_REMOVE",
  TOAST_REMOVE_ALL = "TOAST_REMOVE_ALL",
}

// Admin
export enum USER_ACTIONS {
  USER_LOGIN_REQUEST = "USER_LOGIN_REQUEST",
  USER_LOGIN_SUCCESSED = "USER_LOGIN_SUCCESSED",
  USER_LOGIN_FAILED = "USER_LOGIN_FAILED",
}

export enum MATTERPORT_FRONT_ACTIONS {
  MATTERPORT_ALL_FETCH_REQUESTED = "MATTERPORT_ALL_FETCH_REQUESTED",
  MATTERPORT_ALL_FETCH_SUCCESSED = "MATTERPORT_ALL_FETCH_SUCCESSED",
  MATTERPORT_FETCH_REQUESTED = "MATTERPORT_FETCH_REQUESTED",
  MATTERPORT_FETCH_SUCCESSED = "MATTERPORT_FETCH_SUCCESSED",
  MATTERPORT_FETCH_FAILED = "MATTERPORT_FETCH_FAILED",
}

export enum MATTERPORT_ADMIN_ACTIONS {
  MATTERPORT_SAVE_REQUEST = "MATTERPORT_SAVE_REQUEST",
  MATTERPORT_SAVE_REQUEST_SUCCESSED = "MATTERPORT_SAVE_REQUEST_SUCCESSED",
  MATTERPORT_SAVE_REQUEST_FAILED = "MATTERPORT_SAVE_REQUEST_FAILED",
  MATTERPORT_UPDATE_REQUEST = "MATTERPORT_UPDATE_REQUEST",
  MATTERPORT_UPDATE_REQUEST_SUCCESSED = "MATTERPORT_UPDATE_REQUEST_SUCCESSED",
  MATTERPORT_UPDATE_REQUEST_FAILED = "MATTERPORT_UPDATE_REQUEST_FAILED",
  MATTERPORT_REMOVE_REQUEST = "MATTERPORT_REMOVE_REQUEST",
  MATTERPORT_REMOVE_REQUEST_SUCCESSED = "MATTERPORT_REMOVE_REQUEST_SUCCESSED",
  MATTERPORT_REMOVE_REQUEST_FAILED = "MATTERPORT_REMOVE_REQUEST_FAILED",
}